refactor(types): consolidate Zod schemas as single source of truth
- Move speakerSegmentSummarySchema and extractedSubjectSchema to types.ts
- Use z.infer<> to derive TypeScript types from schemas
- Derive Subject from ExtractedSubject using extends + Omit pattern
- Remove re-exports from aiClient.ts, use direct imports from types.ts
- Explicitly shows transformation: ExtractedSubject → Subject"
